+++ src/List.h
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
+#include <cstdint>
 
 #ifndef List_dot_h
 #define List_dot_h
@@ -8,7 +9,7 @@ class ListMessage {
 public:
 	ListMessage(void);
 	virtual ~ListMessage(void);
-	long type; // application specific poor people's type tag...
+	int32_t type; // application specific poor people's type tag...
 private:
 };
 
@@ -33,7 +34,7 @@ public:
 	virtual void Remove(ListElem *that); // abstract
 	virtual void Broadcast(ListMessage *message, bool forward);
 	static void Delete(ListStruct **list);
-	long elems;			   // to be considered read-only for clients
+	int32_t elems;			   // to be considered read-only for clients
 	ListElem *first,*last; // to be considered read-only for clients
 };
